# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"OrderResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"cart": {},
					"address": {},
					"orderStatus": {
						"type": "string"
					},
					"orderDeliveryStatus": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"cart",
					"address",
					"orderStatus",
					"orderDeliveryStatus",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "core-api",
		"version": "1.4.0",
		"description": "API documentation for the KiranaPro - 'PROD' Backend microservice, designed for seamless integration with multiple microservices, including AI, ONDC/ KIKO/ third-party ONDC services, and mobile frontends (iOS & Android).",
		"license": {
			"name": "Proprietary - Copyright (c) 2024, KiranaPro"
		},
		"contact": {
			"name": "Sandeep K ",
			"email": "sandeep@kirana.pro"
		}
	},
	"paths": {
...
		"/orders": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "A list of orders.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"orders": {
											"items": {
												"$ref": "#/components/schemas/OrderResponse"
											},
											"type": "array"
										}
									},
									"required": [
										"orders"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Fetch all orders for the logged-in user.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "The created order.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"order": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"order",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "CartId and AddressId are required"
					}
				},
				"description": "Create a new order.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The request body containing cartId and addressId.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"addressId": {
										"type": "string"
									},
									"cartId": {
										"type": "string"
									}
								},
								"required": [
									"addressId",
									"cartId"
								],
								"type": "object",
								"description": "The request body containing cartId and addressId."
							}
						}
					}
				}
			}
		},
		"/orders/{id}": {
			"get": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "The order details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderResponse"
								}
							}
						}
					}
				},
				"description": "Get a specific order by ID.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the order to fetch.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/{id}/track": {
			"get": {
				"operationId": "TrackOrder",
				"responses": {
					"200": {
						"description": "The order's tracking status.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deliveryStatus": {
											"type": "string"
										},
										"orderStatus": {
											"type": "string"
										}
									},
									"required": [
										"deliveryStatus",
										"orderStatus"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get the tracking status of a specific order.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the order to track.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/{id}/cancel": {
			"post": {
				"operationId": "CancelOrder",
				"responses": {
					"200": {
						"description": "A success message or an error.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Cancel an order.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the order to cancel.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/v1"
		}
	]
}